split dtls into unsafe code and "safe"
